requirements.txt
pynetworktables
opencv-python
requests